+++ Webpay.Integration.CSharp/Webpay.Integration.CSharp/Order/Row/IRowBuilder.cs
@@ -7,7 +7,7 @@
         string GetUnit();
 
         string GetArticleNumber();
-        int GetDiscountPercent();
+        decimal GetDiscountPercent();
         decimal GetQuantity();
 
         decimal? GetAmountExVat();
